pyproject.toml
[project]
name = "road-network-prolog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geoalchemy2>=0.18.0",
    "networkx>=3.5",
    "osmium>=4.2.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "sqlmodel>=0.0.27",
]
